pyproject.toml
[tool.poetry]
name = "providenie"
version = "0.1.0"
description = "Проект телеграм-бота, который позволяет разгрузить кураторов/координаторов фонда Провидение."
authors = [
    "Карев Виталий", "Дарья Малышева", "Андрей Распутин", "Данилов Николай",
    "Кузнецов Дмитрий", "Моторин Антон", "Курятников Александр",
    "Ирина Финатова", "Юлий Митряшкин", "Яна Денисова"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.0"
python-telegram-bot = "20.0b0"
pydantic = {extras = ["email"], version = "^1.10.6"}
py3-validate-email = "1.0.5.post2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
isort = "^5.11.3"
flake8 = "^6.0.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
